Fixtures moved to file conftest.py
