+++ README.md
@@ -146,7 +146,7 @@
 
 ------------------------------------------------------------------
 
-# Добавленные фикстуры
+# Добавленные фикстуры в модуле *conftest.py*
 
 ## Фикстура *book_and_genre*
